src/routes/requirements.txt
fastapi
uvicorn[standard]
pydantic>=2.0
pydantic-settings
langchain-google-genai
langchain-openai
langchain-community
langchain-ollama
huggingface_hub
browser-use
tqdm
requests
python-multipart
python-dotenv
google-auth
google-auth-oauthlib
playwright
jinja2
itsdangerous
aiohttp
